Module pallet_transfer_allowlist::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

AllowanceDetails

Struct to define when a transfer should be allowed from the sender, receiver, and currency combination. Transfer allowed time set by range of block numbers Defaults to allowed_at starting at 0, and blocked_at ending at MAX block value as per Default impl. Current block must be between allowed at and blocked at for transfer to be approved if allowance for sender/currency/receiver present.

AllowanceMetadata

Metadata values used to track and manage Allowances for a sending Account/Currency combination. contains the number of allowances/presence of existing allowances for said combination, as well as whether a delay is set for the allowance to take effect, and if–and then when–a delay is modifiable.

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Aliases

AccountCurrencyTransferAllowance

Storage item for allowances specified for a sending account, currency type and receiving location

AccountCurrencyTransferCountDelay

Storage item containing number of allowances set, delay for sending account/currency, and block number delay is modifiable at. Contains an instance of AllowanceMetadata with allowance count as u64, current_delay as Option<T::BlockNumber>, and modifiable_at as Option<T::BlockNumber>. If a delay is set, but no allowances have been created, allowance_count will be set to 0. A double map is used here as we need to know whether there is a restriction set for the account and currency in the case where there is no allowance for destination location. Using an StorageNMap would not allow us to look up whether there was a restriction for the sending account and currency, given that:

AllowanceDetailsOf

AllowanceDetails where BlockNumber is of type T::BlockNumber

DepositBalanceOf

Balance type for the reserve/deposit made when creating an Allowance

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.