Type Alias pallet_transfer_allowlist::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn add_transfer_allowance( origin: OriginFor<T>, currency_id: T::CurrencyId, receiver: T::Location ) -> DispatchResult

Adds a transfer allowance for a sending Account/Currency. Allowance either starts at the current block + the delay set for the account, if a delay is present. or block 0 if no delay is present. Important! Account/Currency sets with an allowance set are restricted to just the allowances added for the account - to have unrestricted transfers allowed for the sending Account and Currency, no allowances should be present.

Running this for an existing allowance generates a new allowance based on the current delay, or lack thereof

pub fn remove_transfer_allowance( origin: OriginFor<T>, currency_id: T::CurrencyId, receiver: T::Location ) -> DispatchResult

Restricts a transfer allowance for a sending account/currency/receiver location to:

either the current block + delay if a delay is set

or the current block if no delay is set

pub fn purge_transfer_allowance( origin: OriginFor<T>, currency_id: T::CurrencyId, receiver: T::Location ) -> DispatchResult

Removes a transfer allowance for a sending account/currency and receiving location Decrements or removes the sending account/currency count.

pub fn add_allowance_delay( origin: OriginFor<T>, currency_id: T::CurrencyId, delay: T::BlockNumber ) -> DispatchResult

Adds an account/currency delay Calling on an account/currency with an existing delay will fail. To update a delay the delay has to be set to future modifiable. then an update delay extrinsic called

pub fn update_allowance_delay( origin: OriginFor<T>, currency_id: T::CurrencyId, delay: T::BlockNumber ) -> DispatchResult

Updates an allowance delay, only callable if the delay has been set to allow future modifications and the delay modifiable_at block has been passed.

pub fn toggle_allowance_delay_once_future_modifiable( origin: OriginFor<T>, currency_id: T::CurrencyId ) -> DispatchResult

This allows the delay value to be modified after the current delay has passed since the current block Or sets the delay value to be not modifiable iff modifiable at has already passed

pub fn purge_allowance_delay( origin: OriginFor<T>, currency_id: T::CurrencyId ) -> DispatchResult

Removes an existing sending account/currency delay

impl<T: Config> Pallet<T>

pub fn increment_or_create_allowance_count( account_id: &T::AccountId, currency_id: &T::CurrencyId ) -> DispatchResult

Increments number of allowances present for a sending account/currency set. If no allowances set, an entry with 1 added, if entry already present, it is then incremented.

pub fn decrement_or_remove_allowance_count( account_id: &T::AccountId, currency_id: &T::CurrencyId ) -> DispatchResult

Decrements the number of allowances tracked for a sending account/currency set. If the allowance count is currently 1, then it removes the entry If greater than 1, then decremented. If no entry present, NoAllowancesSet error returned.

impl<T: Config> Pallet<T>

pub fn get_account_currency_restriction_count_delay<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<AllowanceMetadata<T::BlockNumber>>where KArg1: EncodeLike<T::AccountId>, KArg2: EncodeLike<T::CurrencyId>,

Storage item containing number of allowances set, delay for sending account/currency, and block number delay is modifiable at. Contains an instance of AllowanceMetadata with allowance count as u64, current_delay as Option<T::BlockNumber>, and modifiable_at as Option<T::BlockNumber>. If a delay is set, but no allowances have been created, allowance_count will be set to 0. A double map is used here as we need to know whether there is a restriction set for the account and currency in the case where there is no allowance for destination location. Using an StorageNMap would not allow us to look up whether there was a restriction for the sending account and currency, given that:

we’re checking whether there’s an allowance specified for the receiver location

we would only find whether a restriction was set for the account in this case if:

an allowance was specified for the receiving location, which would render blocked restrictions useless

we would otherwise need to store a vec of locations, which is problematic given that there isn’t a set limit on receivers If a transfer restriction is in place, then a second lookup is done on AccountCurrencyAllowances to see if there is an allowance for the receiver This allows us to keep storage map vals to known/bounded sizes.

impl<T: Config> Pallet<T>

pub fn get_account_currency_transfer_allowance<KArg>( key: KArg ) -> Option<AllowanceDetails<T::BlockNumber>>where KArg: EncodeLikeTuple<<(NMapKey<Twox64Concat, T::AccountId>, NMapKey<Twox64Concat, T::CurrencyId>, NMapKey<Blake2_128Concat, T::Location>) as KeyGenerator>::KArg> + TupleToEncodedIter,

Storage item for allowances specified for a sending account, currency type and receiving location